+++ ezmaster-api/routes/v1-hub.js
@@ -1,33 +1,31 @@
 /*eslint-env node */
 /*eslint no-sync: "off", global-require: "off"*/
-'use strict';
+"use strict";
 
-var path = require('path')
-  , basename = path.basename(__filename, '.js')
-  , debug = require('debug')('ezmaster:' + basename)
-  , request = require('request')
-  , url = require('url')
-  ;
+var path = require("path"),
+  basename = path.basename(__filename, ".js"),
+  debug = require("debug")("ezmaster:" + basename),
+  request = require("request"),
+  url = require("url");
 
-
-
-var express = require('express');
+var express = require("express");
 var router = express.Router();
 
-router.route('/*').get(function (req, res, next) {
-  var originalUrl = url.parse(req.originalUrl.replace('/-/v1/hub', '/v2'));
+router.route("/*").get(function(req, res, next) {
+  var originalUrl = url.parse(req.originalUrl.replace("/-/v1/hub", "/v2"));
 
-  originalUrl.protocol = 'http';
-  originalUrl.host = 'hub.docker.com';
+  originalUrl.protocol = "http";
+  originalUrl.host = "hub.docker.com";
 
-  request.get({
-    url : url.format(originalUrl),
-  }).on('error', function (err) {
-    debug('Error when contacting ' + originalUrl.host + ' - ' + err);
-    return res.status(500).send(err);
-  }).pipe(res);
-
+  request
+    .get({
+      url: url.format(originalUrl)
+    })
+    .on("error", function(err) {
+      debug("Error when contacting " + originalUrl.host + " - " + err);
+      return res.status(500).send(err);
+    })
+    .pipe(res);
 });
-
 
 module.exports = router;