+++ ezmaster-front/src/tab-applications/ApplicationRow.js
@@ -30,6 +30,8 @@ class ApplicationRow extends Component {
             application={this.props.application}
             applications={this.props.applications}
             usedApplications={this.props.usedApplications}
+            className="ezmaster-a-a"
+            classNameBtn="ezmaster-a-a-btn"
           />
         </td>
       </tr>